docs: clarify security defaults for ClawScan findings, move schema to internal
